+++ BlueprintUI/Sources/Layout/Empty.swift
@@ -1,0 +1,25 @@
+//
+//  Empty.swift
+//  BlueprintUI
+//
+//  Created by Kyle Van Essen on 6/4/20.
+//
+
+import UIKit
+
+
+///
+/// An empty `Element` which has no size and draws no content.
+///
+public struct Empty : Element {
+    
+    public init() {}
+    
+    public var content: ElementContent {
+        ElementContent(intrinsicSize: .zero)
+    }
+    
+    public func backingViewDescription(bounds: CGRect, subtreeExtent: CGRect?) -> ViewDescription? {
+        nil
+    }
+}